requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pydantic-settings
pyodbc
python-jose
pydantic[email]
python-multipart
sqlalchemy
passlib[bcrypt]